Fix building and linting
