--- client/org/dst/client/DstClient.java
@@ -1,12 +1,0 @@
-package org.dst.client;
-
-public interface DstClient {
-
-    boolean connect();
-
-    boolean isConnected();
-
-    boolean disconnect();
-
-    DstStringProxy str();
-}
