--- client/org/dst/client/exception/KeyNotFoundException.java
@@ -1,8 +1,0 @@
-package org.dst.client.exception;
-
-public class KeyNotFoundException extends DstException {
-
-    public KeyNotFoundException(String key) {
-        super(String.format("The key %s doesn't exist in the store.", key));
-    }
-}
