+++ client/src/main/java/org/dst/client/DstClient.java
@@ -1,0 +1,12 @@
+package org.dst.client;
+
+public interface DstClient {
+
+  boolean connect();
+
+  boolean isConnected();
+
+  boolean disconnect();
+
+  DstStringProxy str();
+}
