+++ client/src/main/java/org/dst/client/exception/DstException.java
@@ -1,0 +1,13 @@
+package org.dst.client.exception;
+
+public class DstException extends RuntimeException {
+
+  public DstException(String errorMessage) {
+    super(errorMessage);
+  }
+
+  public DstException(String errorMessage, Exception e) {
+    super(errorMessage, e);
+  }
+
+}
